+++ src/main/java/com/kdb/it/domain/budget/work/repository/BudgetReadView.java
@@ -16,5 +16,5 @@ public interface BudgetReadView {
 
     BigDecimal getBgDupAmt();
 
-    Integer getAsgRt();
+    BigDecimal getAsgRt();
 }
